@@ -1,71 +1,70 @@
 {
   "name": "pip-webui2-layouts",
-  "version": "1.1.21",
+  "version": "2.0.0-rc.1",
   "license": "MIT",
   "scripts": {
     "ng": "ng",
-    "start": "ng serve",
-    "start_prod": "ng serve --prod",
-    "build": "ng build",
-    "build_prod": "ng build --prod",
-    "build:lib": "ng build pip-webui2-layouts && npm run build:lib-styles",
+    "start": "pip-webui2-themes build --soft && ng serve",
+    "build": "pip-webui2-themes build && ng build",
+    "build:lib": "ng build pip-webui2-layouts --prod && npm run build:lib-styles && npm run build:lib-md",
     "build:lib-styles": "scss-bundle -c ./projects/pip-webui2-layouts/scss-bundle.config.json",
+    "build:lib-md": "cpx \"{LICENSE,README.md,CHANGELOG.md}\" dist/pip-webui2-layouts",
     "test": "ng test",
     "test:lib": "ng test pip-webui2-layouts",
     "lint": "ng lint",
     "lint:lib": "ng lint pip-webui2-layouts",
     "e2e": "ng e2e",
-    "package": "npm run build:lib && (for %I in (LICENSE README.md CHANGELOG.md) do copy .\\%I .\\dist\\pip-webui2-layouts) && cd dist/pip-webui2-layouts && npm pack",
+    "package": "npm run build:lib && cd dist/pip-webui2-layouts && npm pack",
     "deploy": "ng build --prod --base-href https://pip-webui2.github.io/pip-webui2-layouts/ && ngh --dir dist/pip-webui2-layouts-app"
   },
   "private": false,
   "dependencies": {
-    "@angular/animations": "^7.2.0",
-    "@angular/cdk": "^7.2.1",
-    "@angular/common": "^7.2.0",
-    "@angular/compiler": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/forms": "^7.2.0",
-    "@angular/http": "^7.2.0",
-    "@angular/material": "^7.2.1",
-    "@angular/platform-browser": "^7.2.0",
-    "@angular/platform-browser-dynamic": "^7.2.0",
-    "@angular/router": "^7.2.0",
-    "@ngx-translate/core": "^11.0.1",
-    "core-js": "^2.6.2",
+    "@angular/animations": "^9.0.4",
+    "@angular/cdk": "^9.1.0",
+    "@angular/common": "^9.0.4",
+    "@angular/compiler": "^9.0.4",
+    "@angular/core": "^9.0.4",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/forms": "^9.0.4",
+    "@angular/material": "^9.1.0",
+    "@angular/platform-browser": "^9.0.4",
+    "@angular/platform-browser-dynamic": "^9.0.4",
+    "@angular/router": "^9.0.4",
+    "@ngx-translate/core": "^12.1.2",
+    "core-js": "^3.6.4",
     "lodash": "^4.17.11",
     "masonry-layout": "^4.2.2",
-    "node-sass": "^4.11.0",
-    "pip-webui2-themes": "^1.1.24",
-    "rxjs": "~6.3.3",
+    "node-sass": "^4.13.1",
+    "pip-webui2-themes": "^2.0.0-rc.2",
+    "rxjs": "~6.5.4",
     "scss-bundle": "^2.4.0",
-    "zone.js": "~0.8.27"
+    "tslib": "^1.10.0",
+    "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.12.1",
-    "@angular-devkit/build-ng-packagr": "~0.12.1",
-    "@angular/cli": "^7.2.1",
-    "@angular/compiler-cli": "^7.2.0",
-    "@angular/language-service": "^7.2.0",
-    "@types/jasmine": "~3.3.5",
+    "@angular-devkit/build-angular": "~0.900.3",
+    "@angular-devkit/build-ng-packagr": "~0.900.3",
+    "@angular/cli": "^9.0.3",
+    "@angular/compiler-cli": "^9.0.2",
+    "@angular/language-service": "^9.0.2",
+    "@types/jasmine": "~3.3.0",
     "@types/jasminewd2": "~2.0.6",
-    "@types/node": "~10.12.18",
-    "angular-cli-ghpages": "^0.5.3",
-    "codelyzer": "~4.5.0",
+    "@types/lodash": "^4.14.149",
+    "@types/node": "^12.11.1",
+    "angular-cli-ghpages": "^0.6.2",
+    "codelyzer": "^5.1.2",
+    "cpx": "^1.5.0",
     "jasmine-core": "~3.3.0",
     "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~3.1.4",
+    "karma": "~4.4.1",
     "karma-chrome-launcher": "~2.2.0",
     "karma-coverage-istanbul-reporter": "~2.0.1",
     "karma-jasmine": "~2.0.1",
     "karma-jasmine-html-reporter": "^1.4.0",
-    "ng-packagr": "^4.6.0",
-    "protractor": "~5.4.2",
-    "ts-node": "~7.0.1",
-    "tsickle": "0.34.0",
-    "tslib": "^1.9.0",
-    "tslint": "~5.12.1",
-    "typescript": "~3.2.2"
+    "ng-packagr": "^9.0.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.11.0",
+    "typescript": "~3.7.5"
   }
 }
