@@ -1,22 +1,21 @@
 {
   "name": "pip-webui2-layouts",
-  "version": "1.1.21",
+  "version": "2.0.0-rc.1",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "https://github.com/pip-webui2/pip-webui2-layouts"
   },
   "dependencies": {
-    "@ngx-translate/core": "^11.0.1",
     "masonry-layout": "^4.2.2",
-    "pip-webui2-themes": "^1.1.24"
+    "pip-webui2-themes": "^2.0.0-rc.2"
   },
   "peerDependencies": {
-    "@angular/cdk": "^7.2.0",
-    "@angular/common": "^7.2.0",
-    "@angular/core": "^7.2.0",
-    "@angular/flex-layout": "^7.0.0-beta.23",
-    "@angular/material": "^7.2.1",
-    "rxjs": "~6.3.3"
+    "@angular/cdk": "^9.1.0",
+    "@angular/common": "^9.0.4",
+    "@angular/core": "^9.0.4",
+    "@angular/flex-layout": "^9.0.0-beta.29",
+    "@angular/material": "^9.1.0",
+    "rxjs": "~6.5.4"
   }
 }
