@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatSidenavModule } from '@angular/material';
+import { MatSidenavModule } from '@angular/material/sidenav';
 
 import { PipSidenavService } from './shared/sidenav.service';
 import { ModuleWithProviders } from '@angular/compiler/src/core';
